+++ modules/ecs_service_fargate/variables.tf
@@ -74,4 +74,3 @@ variable "containers_per_task" {
   description = "Number of container to run per task. Used to specify the amount of memory for each container if running multiple per task."
   default     = 1
 }
-